requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
google-generativeai>=0.3.0
pandas>=2.0.0
streamlit-chat>=0.1.0